Move shared preference logic into MainActivity and fix type of NightScoutData
