+++ wear/src/main/java/com/kaist/k_dual/model/UseSetting.kt
@@ -18,11 +18,7 @@ object UseSetting {
     private val gson = Gson()
     var settings: Setting? by mutableStateOf(null)
 
-    private val preferenceChangeListener = SharedPreferences.OnSharedPreferenceChangeListener { _, _ ->
-        updateSettings()
-    }
-
-    private fun updateSettings() {
+    fun updateSettings() {
         val jsonString = sharedPreferences.getString(SETTINGS_KEY, null)
         Log.d("updateSettings", jsonString ?: "")
         settings = if (jsonString != null) {
@@ -41,13 +37,5 @@ object UseSetting {
     fun initialize(context: Context) {
         sharedPreferences = context.getSharedPreferences(PREFERENCES_FILE_KEY, Context.MODE_PRIVATE)
         updateSettings()
-    }
-
-    fun registerListener() {
-        sharedPreferences.registerOnSharedPreferenceChangeListener(preferenceChangeListener)
-    }
-
-    fun unregisterListener() {
-        sharedPreferences.unregisterOnSharedPreferenceChangeListener(preferenceChangeListener)
     }
 }
