+++ wear/src/main/java/com/kaist/k_dual/presentation/MainActivity.kt
@@ -1,6 +1,9 @@
 package com.kaist.k_dual.presentation
 
+import android.content.Context
 import android.content.Intent
+import android.content.SharedPreferences
+import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.Box
@@ -14,14 +17,28 @@ import androidx.navigation.navArgument
 import androidx.wear.compose.navigation.SwipeDismissableNavHost
 import androidx.wear.compose.navigation.composable
 import androidx.wear.compose.navigation.rememberSwipeDismissableNavController
+import com.kaist.k_canvas.PREFERENCES_FILE_KEY
+import com.kaist.k_dual.model.UseBloodGlucose
 import com.kaist.k_dual.model.UseSetting
 import com.kaist.k_dual.presentation.theme.KDualTheme
 import com.kaist.k_dual.service.GlucoseUpdateService
 
 class MainActivity : ComponentActivity() {
-    override fun onStart() {
-        super.onStart()
+    private lateinit var sharedPreferences: SharedPreferences
+    private val preferenceChangeListener = SharedPreferences.OnSharedPreferenceChangeListener { _, _ ->
+        UseSetting.updateSettings()
+        UseBloodGlucose.updateBloodGlucose(this)
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        sharedPreferences = getSharedPreferences(PREFERENCES_FILE_KEY, Context.MODE_PRIVATE)
+
         UseSetting.initialize(this)
+
+        sharedPreferences.registerOnSharedPreferenceChangeListener(preferenceChangeListener)
+
         startService(Intent(this, GlucoseUpdateService::class.java))
         val userId = intent.getIntExtra("routeGraph", 0)
         setContent {
@@ -31,12 +48,17 @@ class MainActivity : ComponentActivity() {
 
     override fun onResume() {
         super.onResume()
-        UseSetting.registerListener()
+        sharedPreferences.registerOnSharedPreferenceChangeListener(preferenceChangeListener)
     }
 
     override fun onPause() {
         super.onPause()
-        UseSetting.unregisterListener()
+        sharedPreferences.unregisterOnSharedPreferenceChangeListener(preferenceChangeListener)
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        sharedPreferences.unregisterOnSharedPreferenceChangeListener(preferenceChangeListener)
     }
 }
