+++ wear/src/main/java/com/kaist/k_dual/service/ListenerService.kt
@@ -1,4 +1,4 @@
-package com.kaist.k_dual.listener
+package com.kaist.k_dual.service
 
 import android.content.Context
 import android.util.Log
